aulataller/requirements.txt
streamlit
ultralytics
opencv-python
pillow
streamlit-image-comparison